requirements.txt
streamlit==1.28.1
pillow==10.0.1
pandas==2.1.1
openpyxl==3.1.2
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0